Kernel revamp

+++ include/i386/regs.h
@@ -1,7 +1,9 @@
+#include <stdint.h>
+
 struct regs
 {
-    unsigned int gs, fs, es, ds;
-    unsigned int edi, esi, ebp, esp, ebx, edx, ecx, eax;
-    unsigned int int_no, err_code;
-    unsigned int eip, cs, eflags, useresp, ss;
+    uint16_t gs, fs, es, ds;
+    uint16_t edi, esi, ebp, esp, ebx, edx, ecx, eax;
+    uint16_t int_no, err_code;
+    uint16_t eip, cs, eflags, useresp, ss;
 };
